% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeOD.R
\name{changeOD}
\alias{changeOD}
\title{Explore changes in OD datasets over time}
\usage{
changeOD(x, y)
}
\arguments{
\item{x}{Origin-destination dataset 1}

\item{y}{Origin-destination dataset 2}
}
\value{
A plot comparing the two datasets
}
\description{
Explore changes in OD datasets over time
}
\examples{
\dontrun{
changeOD(x = od1, y = od2)
}
}
